fs_mgr: When formating f2fs volumes, respect the length parameter
When formatting a volume because it is marked as formattable, the
length parameter is discarded which makes fs_mgr write a filesystem
to the full length of the block device.

This patch adds length semantics to f2fs formatting, if the length
is greater than zero, use that, if it isn't subtract that size
from the block size.

Change-Id: I526f80aa299e7b34e9802141e7fa7050d5cb5558
Ticket: SAMBAR-729
(cherry picked from commit 8c48aafb7dd7aca4086a6bbc1a97d4fcc49ff98c)