Add vendor hook to handle_control_message
This is needed for Nvidia Shield devices to handle a
'restart consolemode' request from a blobbed stock app.

The vendor function is expected to return non-0 if it did
not handle the message and 0 if it did (or wants it ignored).

Forward port of 119414.

Change-Id: Ia8e4ba0bbf561f29f72862cd986f1660d7b501da
(cherry picked from commit 85de7b041c010f174d8df3cb2976a93ce7b252cb)